fix various list issues
